+++ changelog.md
@@ -2,3 +2,7 @@
 
 - ***Added*** carpet rule `betterMobCap` to limit phantom and pillager patrol spawning
 - Updated translations
+
+### Bugfix
+
+- Unmatched behaviour of `loadedChunks` logger in different Minecraft versions